Commit 3

+++ src/Controller/PanierController.php
@@ -5,12 +5,16 @@ namespace App\Controller;
 use App\Entity\Discount;
 use App\Entity\Product;
 use App\Entity\ProductQuantity;
+use App\Entity\Stripe;
 use App\Entity\User;
+use App\Service\CheckoutSession;
+use Stripe\Checkout\Session;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 class PanierController extends AbstractController
 {
@@ -35,36 +39,87 @@ class PanierController extends AbstractController
         ]);
     }
 
-    #[Route('/checkout/{promo}', name: 'checkout')]
+    #[Route('/checkout', name: 'checkout')]
     public function checkout(): Response
     {
         $em = $this->getDoctrine()->getManager();
         $user = $em->getRepository(User::class)->find($this->getUser()->getId());
         $quantities = $em->getRepository(ProductQuantity::class)->findBy(['Panier' => $user->getPanier()]);
 
-        return $this->render('panier/checkout.html.twig', [
+        return $this->render('panier/checkout/checkout.html.twig', [
             'quantities' => $quantities,
             'user' => $user,
         ]);
     }
 
-    #[Route('/pay', name: 'pay')]
-    public function pay(): Response
+    #[Route('/checkout/session', name: 'set_checkout')]
+    public function setCheckout(CheckoutSession $service)
     {
         $em = $this->getDoctrine()->getManager();
         $user = $em->getRepository(User::class)->find($this->getUser()->getId());
         $quantities = $em->getRepository(ProductQuantity::class)->findBy(['Panier' => $user->getPanier()]);
 
-        foreach ($quantities as $quantity) {
-            $quantity->getPanier()->setTotalPrice($quantity->getPanier()->getTotalPrice() - $quantity->getTotalQuantityPrice());
-            $em->remove($quantity);
-        };
+        $products = [];
 
-        $em->flush();
+        foreach($quantities as $quantity)
+        {
+            $name = strval($quantity->getProduct()->getName());
+            // $productQuantity = strval($quantity->getProduct()->getProductQuantities());
+            $product = [
+                'price_data' => [
+                  'currency' => 'eur',
+                  'unit_amount' => $quantity->getProduct()->getPrice()*100,
+                  'product_data' => [
+                    // 'name' => $name,
+                    'name' => $quantity->getProduct()->getName(),
+                  ],
+                ],
+                // 'quantity' => 2,
+                'quantity' => $quantity->getQuantity(),
+              ];
+            array_push($products, $product);
+        }
+        $sucessUrl = $this->generateUrl('success', [], UrlGeneratorInterface::ABSOLUTE_URL);
+        $id = $service->createCheckoutSession($products, $sucessUrl);
+        return new JsonResponse(['id' => $id]);
+    }
 
-        return $this->render('panier/pay.html.twig', [
-            'user' => $user,
-        ]);
+    #[Route('/success', name: 'success')]
+    public function sucess(Request $request, CheckoutSession $service): Response
+    {
+        $em = $this->getDoctrine()->getManager();
+        $user = $em->getRepository(User::class)->find($this->getUser()->getId());
+        $quantities = $em->getRepository(ProductQuantity::class)->findBy(['Panier' => $user->getPanier()]);
+        $session = $service->getSessionById($request->get('session_id'));
+        $paid = $session->payment_status;
+        // dd($session);
+        if($session instanceof Session){
+            $stripe = new Stripe();
+            $stripe->setSessionId($request->get('session_id'));
+            if($paid === 'paid'){
+                $result = $this->render('panier/checkout/success.html.twig', []);
+                $stripe->setPaymentStatus($paid);
+                foreach($quantities as $quantity){
+                    $em->remove($quantity);
+                }
+            }else{
+                $result = $this->render('panier/checkout/failed.html.twig', []);
+                $stripe->setPaymentStatus($paid);
+            }
+            $em->persist($stripe);
+            $em->flush();
+        }else{
+            throw $this->createNotFoundException('not_exist');
+        }
+        return $result;
+    }
+
+    #[Route('/cancel', name: 'cancel')]
+    public function cancel(): Response
+    {
+
+
+        return $this->render('panier/checkout/cancel.html.twig', []);
     }
 
     #[Route('/add_to_cart', name: 'add_to_cart')]
@@ -137,7 +192,7 @@ class PanierController extends AbstractController
         $promo = '2';
         if ($discount && $discount->getUseCount() >= 1 && $user->getPanier()->getDiscountUsed() === false) {
             $user->getPanier()->setTotalPrice($user->getPanier()->getTotalPrice() - $user->getPanier()->getTotalPrice() * $discount->getDiscount() / 100)->setDiscountUsed(true);
-            $discount->setUseCount($discount->getUseCount()-1);
+            $discount->setUseCount($discount->getUseCount() - 1);
             $em->flush();
             $promo = '1';
         } elseif ($discount && $user->getPanier()->getDiscountUsed()) {
